UUIDSerializer: Change type cast to parent type to fix issues using within a list.
The UUIDSerializer's deserializer will fail with a `ClassCastException` when the UUID is embedded in a list (e.g., `val foo: Collection<@Contextual UUID> = emptyList()`).  The decoder in this example is a `ListDecoder` rather than the expected `BsonFlexibleDecoder` triggering the exception.  Both, however, are `FlexibleDecoder` objects and the reader field is declared there.  Updated the cast to this parent type, resolving the issue.